refactor: migrate convert-lesson-image to Payload endpoint
- Move from /api/ai/convert-lesson-image (Next.js route) to
  /api/lessons/:id/import-exercise (Payload endpoint)
- Use req.user for native authentication (no manual auth)
- Use req.payload for direct Payload SDK access
- Resource-based URL structure follows REST conventions
- Remove old Next.js route file entirely
- Benefits: cleaner architecture, consistent access control
